requirements.txt
python-decouple==3.8
python-dotenv==1.1.0
